[Refactor #435]: Move state up the component
