+++ app/components/live.hbs
@@ -1,0 +1,6 @@
+<LiveHeader
+  @tabs={{this.TABS}}
+  @activeTab={{this.activeTab}}
+  @tabHandler={{this.tabHandler}}
+/>
+<p>{{this.activeTab}} Component</p>
