+++ app/templates/live.hbs
@@ -1,1 +1,1 @@
-<LiveHeader/>
+<Live/>
